Try to be more consistent with rationalness of timestamp.
Now, every event coming from users or admins (i.e. questions,
submissions, announcements, ...) is saved as an integer, and in
subsequent computations it is assumed to be an integer. We use
floating point timestamps only when necessary (i.e. when triggering
timeouts in async, or computing resources usage in an amount of time).

Note that db already stored timestamps of the first kind as an
integer.