Corrected bugs in rpc via ajax, and add authorizations.
Now rpc methods that the browser can call needs to be allowed in the
WS, in the method "authorized_rpc". This can obviously check things
like "the user is asking for information about a different user?".

The bug was something related to closures. God, I hate closures. Just
joking, I love them. For some reasons the bind was wrong and when I
got two requests in the same page, the response was always for the
last request rid. Now it should be fixed.