First try for accepting >1 files in a submission.
Also slightly changed the semantic in Files: now the names are the
same as the one in submission_format, and the submission has an
additional field "language", that specify how to translate the %l
found in the filenames.

Another change is that the formal correctedness (and language
infering) of the submission is done by CWS at the time of the
submission, and invalid ones are not accepted and not stored.