Using bigger timeout for Async services (not Tornado) - see after the cut.
Ok here's the thing. Standard practice asks not to merge select loops
(as is asyncore's loop) and timeouts. But we do it anyway, hey, no
problems. Yet, it is good to make the select as long as possible (ever
heard of busy wait?).

Consider now that in a Async-only service (i.e. not webservice) we do
three kind of things when the select ends:
1. try to connect to disconnected services (very low priority);
2. send reply for threaded RPC that ended (middle priority, given that
   we use them only to do worker jobs);
3. check if a timeout as expired and execute it (high priority).

The first may as well be executed every bunch of seconds, we don't
care a lot. The second maybe not less than every couple of
seconds. The third, hell, we know when we need it!

Hence we may set the select timeout as the minimum between the
expiration of the timeouts and 2 seconds (to account for 1. and 2.).

But, it turns out that mixing Tornado's and asyncore's loops was not
such a good idea (TM). At least the way we do it. If we let asyncore
selecting for a couple of seconds, Tornado is unresponsive in that
time and cannot take request from the web.

Hence, we leave as a temporary fix the maximum select time of 0.02
seconds for webservices. Yay, life sucks and we need to find a better
way to handle webservices.