Improved ResourceService.
First of all fix the interaction with psutil (now should work with
every version >= 0.2.0).

More important, with the -a switch, ResourceService is intructed to
autostart all services in its machine, and restart them if they
die. The only services that are left to the user are RS themselves (of
course) and LogService.