SS checks periodically if some submission has slipped without score.
This was a little harder than in ES because SS needs anyway to load
all submissions at start, so the equation "submissions to load ==
submissions without a score" does not hold. Besides, we don't want to
nullify all scores each time we restart SS, because for a while
perfectly valid scores will be unavailable.

My solution was (as ugly as it seems) to add to SS a set of "examined
submissions". The periodic check goes then from linear (as in ES) to
nlogn, but this should be manageable anyway since it is executed only
sporadically.