Finished rewriting the ranking client in pure JavaScript.
The class structure is almost identical to the one we had with Python.
A few external libraries had to be included, most notably JQuery (and a
couple of polyfills). Dummy images for the user photo and team flag were
added. The setup was updated to remove any reference to pyjamas.