Use built-in binary search in ResourceService
We discovered that the binary search implementation in ResourceService
(used to ensure only recent data is sent to the client) sometimes
returns incorrect results. This leads to large amounts of data being
repeatedly being sent to the client.

While this bug has since been fixed in mainline, a better approach may
be to use Python's in-built "bisect" module which implements the same
algorithm.