Do not install certain files if the destinatio is a symlink.
In some cases it is useful to substitute a few CMS files and directories
with symbolic links, for example pointing to a common /data dir.
If setup.py detects that some files being overwritten are symlinks,
it does ignore them (the system administrator will have to ensure
that they are up-to-date and with correct permissions).