Add support for ignoring contestant questions on the admin interface.
On the admin interface questions are prominently displayed until a reply
has been issued. Often, however, questions do not need a reply; in this
case, it is useful to mark a question as "ignored" and hidden.

This patch adds a new boolean column "ignored" to the questions table
and adds a web interface allowing questions to be marked as "ignored"
(or, later, "unignored"). Ignored questions are not counted on the
unanswered message count and appear faded on the questions display
screen.