Ensure questions submitted in the same second are displayed in deterministic order.
Questions are currently displayed in the admin panel from most recent to
least recent based on their timestamps. Unfortunately, if two questions
are submitted in the same second, their order may change each time
a page is loaded.

We change the query to perform a secondary sort by question ID, ensuring
that the output order is deterministic.