async_lock only used in async.
These apparently solves a mysterious bug where the worker hung for
minutes before acknowledging a message from ES. The bug happened only
on virtual machines (because of single-threadness?).

It is also much more sane to code and seems to solve the problem
exactly in the same way as before, but in a more consistent way
(coders don't need to remember to aquire the lock before an RPC).