Make RWS reject requests that would make it inconsistent
This includes:
* creating or updating a task with a non-existing contest
* creating or updating a user with a non-existing team
* creating or updating a sumbission with a non-existing task or user
* creating or updating a subchange with a non-existing submission

In addition, deleting an entity will cause all entities referring to it
to be deleted too.

To accomplish these two goals we needed circular imports. The only way
we found to make it work was to use local imports (i.e. `import Foo')
instead of global imports (i.e. `import cmsranking.Foo as Foo'). Any
suggestion on how to use global imports is welcome.

There are also a couple of bug fixes in Scoring.py.