AWS asks SS to resend data to RWS when some relevant data changes.
This is preliminary support, and does not handle deletion of things
(well, AWS doesn't too), in the sense that, RWS keeps displaying
them. More important, if a task or a user changes name, RWS create a
new task/user and treats the two as different. Need to talk with Luca
on how to solve this.

It seems safer anyway to act more directly for an important contest
(i.e., stopping SS and RWS, clearing RWS data, restarting both).