Reenabled threaded on precache, and don't cache executables.
I couldn't find problems with rpc_threaded precache. In case, we can
comment it out again.

As for not caching executables, it is "necessary" for long contests
(e.g., months) where the number of executables can be huge and the
admin does not want to waste GBs of storage for nothing. (Without
precaching them, it is possible to delete the cache and the worker
won't download them again.)