Added a new scoretype and a new tasktype to handle the task saveit (IOI2010).
We added a new task type TwoSteps for task in the format
"encoder/decoder": the user must implement two functions that get
compiled together with a manager. One function gets called, and return
some data that is passed to the second function.

There is also a score type Threshold, where the score of a group is
awarded if all outcomes are less or equal a specific threshold for
that group.