Use more standard constructions in the Cryptographics module
In particular:
 - use a random IV instead of a fixed IV and a "random salt"
 - use a padding scheme that allows trailing null bytes in the plaintext
   (causing a small increase in the ciphertext size in certain cases)

These changes don't affect the security of the cryptographic functions.
Yet, even if I am convinced that these functions are secure against
eavesdropping, I'm not sure about their security against active attacks
(that is, tampering, replay- and timing-attacks). We'll need to check
that in more detail, sooner or later.

Also, this commit fixes an error that happens with pycrypto-2.6, which
changed its API by making the IV parameter mandatory.