Reorganization of score types.
The details returned by the score type are no longer additional data,
but the whole table of details of the submission. RWS must be informed
of this change, since now it expects details as a list of strings
(JSON encoded).

Moreover, a lot of common code for Group score types has been factored
out in ScoreType.py.