Allow tokens to be infinite
Change some of the constraints about token parameters, and also allow
tokens to be infinite. In particular:
- avoid token_max and token_total being zero (the "behavior" from the
  user's perspective is the same as disabling tokens completely);
- avoid token_min_interval being None and allow it being zero;
- avoid token_gen_time and token_gen_number being None and allow them
  being zero.

The semantic of token_gen_time and token_gen_number slightly changes:
- if _gen_number == 0: don't generate tokens
- elif _gen_time == 0: give infinite tokens
- else: generate _gen_number tokens every _gen_time minutes

At the moment CWS doesn't show correctly to the user that he/she has
infinite tokens. It will be fixed soon. AWS, on the other hand, does
not need any fix since it just shows the parameters' values without
any interpretation.

This commit also fixes a couple of bugs in the token generation
function.