Make the DateTime fields in the DB be in UTC instead of local timezone
This way we have no ambiguity on DST boundaries and it's easier to do
timezone conversions (we'll need this when we make CWS timezone-aware).

Also, the format in which datetimes are displayed in CWS is now
localizable. On the other hand, AWS displays (and requires as input)
datetimes in UTC in ISO format (with a space as separator).