Modify the way CWS stores a local copy of the submissions
CWS was using pickle.dump to store a local copy of the submissions, to
recover them in case of failure. It was dumping the ID of the Contest,
the ID of the User and the Task object, along with all the submitted
files. Now CWS stores just the ID of the Task (instead of the whole
Task).

Storing the whole Task object makes pickle fail because of some circular
references among objects, i.e. task == task.contest.tasks[i] for some i.
Pickle should handle them correctly (in my opinion), yet some bugs are
present[1].

This bug has been introduced by commit c7eb5bb. It wasn't present before
because, thanks to the lazy-loading of SQLALchemy relationships[2], the
contest.tasks attribute was never requested and, thus, never loaded.
This prevented pickle from trying to dump it and falling into infinite
recursion.

Fixes issue #16 (see that issue for more details on this bug).

[1] http://bugs.python.org/issue9269 (note that mike bayer, aka zzzeek,
    commenting on this issue is the author of SQLAlchemy)
[2] http://docs.sqlalchemy.org/en/rel_0_7/orm/loading.html