Show (dates and) times in CWS using a per-user and per-contest timezone
Format the timestamps using the per-user timezone or (if None or
invalid) the per-contest timezone or (if None or invalid) the local
timezone of the server.

All dates and times shown in CWS should be using this timezone. At the
moment the "server time" is an exception: it's shown using the user's
local timezone (as seen by JavaScript, thus the browser). Maybe this is
what we want?

The timezone is currently used only to format timestamps. That is, it
can't be used in an APIO-like setting where the contest start and stop
time depends on the user's timezone.