Fix bug that prevented ES from communicating with the workers
The bug is caused by the fact that ES is sending a datetime object
inside the "plus" argument of an asynchornous RPC call to the Worker,
that AsyncLibrary uses JSON to encode the arguments of these calls and
that JSON doesn't support decoding/encoding of datetime objects.

The only possible solution (except using another serialization format,
like YAML or Pickle) is to transform the datetimes in floating point
timestamps before calling the method, and transform them back as the
method returns.

Other fixes include using "%s" instead of "%d" when formatting
timedeltas. This automatically produces strings like "1 day, 01:23:45".