Add better support for translations of the task statement
Add a languages field to each user to store a list of languages. These
languages are used to present (i.e. highlight) some suggested
translation of the task statements (if these translations exist).
The list is also used to select the locale used to translate the
interface for the user.

If the languages are in the form "%l" or "%l_%c" (with %l being an ISO
639 alpha-2 code and %c being an ISO 3166 alpha-2 code) then their name
is expanded into a human-readable localized form when shown in CWS.

There's currently no support in AWS to view or edit the languages field.