Switch Worker to the Job-based compilation system.
This stage is temporary: at the moment, the Jobs are created by the Worker,
that implements towards the EvaluationService the same protocol as before.

Instead, the Job should be instantiated and sent to the Worker directly by
the EvaluationService, that then should interpret the responses in a way
similar to how it does right now.

Anyway, I'm doing these modifications a bit a time, to better assess their
impact before doing the whole work.