Fix some "User %s called with undecryptable argument." warnings in CWS
They were caused by the JS code in the task submissions page used to
auto-update the submission status. It was trying to do that even when
there were no submissions. Yet, in this case the only row in the table
is the "no submissions yet" row, which hasn't a data-submission
attribute. The JS code tried to get that attribute to make the AJAX
request, obtaining an URL like "/submission_status/undefined", which
obviously failed.