Use cleaner URLs in CWS
Do not encrypt parts of the URL but, instead, use the task name in place
of the task id, a progressive number in place of the submission id and
the file name in place of the the file id. This achieves the same effect
we used cryptography for in the first place: to avoid users seeing
internal ids and using them, for example, to obtain the total number of
submissions received so far.

We have to quote all these values in the URLs to avoid ambiguities. That
is, a URL in the form '%s/bar/%s' % ('foo/bar', 'bar/baz') cannot be
unambiguously split into the original tokens unless the slashes are
escaped. In addition, the parsing may not work if these tokens contain
characters like '?', '#', etc. (See also RFC 3986, section 2 [1])

[1] http://tools.ietf.org/html/rfc3986#section-2