Improve handling of MIME types in CWS
Move the MIME-type-detection functionalities to cmscommon.MimeTypes.
At the moment it's just a small wrapper on the Python mimetypes module,
which enhanced support for the most commonly used extensions (in out
context). In the future me way want to switch to a better library
(perhaps based on shared-mime-info) and provide also detection based on
content rather than on name. This commit makes it easier to do such
backend changes.

In addition, avoid failing when we don't recognize the MIME type (by
using a default of 'application/octet-stream').

Also, detect and provide MIME types for submission files, instead of
using always 'text/plain'.