Show time and memory consumptions in the submission details in CWS
In the submission details window in CWS show two additional columns with
the time and memory consumption of the submission for each testcase.
When these data isn't available a 'N/A' is shown. When the task doesn't
have time or memory limit the corresponding column isn't shown (this
happens mainly for output-only tasks).