Some fixes to Server-Sent-Events communication (RWS server side)
Use the X-Requested-With header to distinguish requests made by the
EventSource polyfill from the one made natively by the browser. In the
former case we need to immediately finish the request as soon as we send
a message (since otherwise some browser won't dispatch the event).
This requires us to listen to on_finish to properly clean up everything.

In addition, make all requests non-cachable. This is because if we
restart the server without a change in the data then the clients may
use the cached data, with the cached (old) timestamp. They'll think
(erroneously) that their data is outdated but even by refreshing they
won't be able to get an updated one.