Small style changes for RWS
Make .inner_frame scrollable instead of the body. This removes the need
to hide the scrollbar when a panel (UserDetail or TeamSearch) is opened,
which caused a lot of redraw (and pehaps even reflow), which was
terribly slow. Also, the scrollbar is now below the upper panel, instead
of on its right side.

Also, remove the top and right borders from the upper panel (by setting
its width to 0), instead of showing them with a width of 2px but then
"hiding" them by moving the panel -2px to top and left. The same applies
to the left border of the side panel.

Align the flag images vertically inside their table cell.