Allow HTTPS between SS and RWS
SS used to send username and password in plaintext in its requests to
RWS, allowing an eavesdropper to get them and send custom forged
requests to RWS. To fix this and to keep using HTTP Basic authentication
a transport layer encryption such as HTTPS is the ideal solution.

HTTPS is optional both in SS and in RWS. Plain HTTP communication can
still be used and username and password will still be sent in plaintext,
but this is intended mainly for cases where RWSs are used privately. For
public RWSs it's reccomended to use HTTPS.

When using HTTPS, SS will send username and password only if the HTTPS
certificate provided by RWS is trusted (it's checked using SS's own list
of trusted certificates). Thus, the communication will happen only if
both parties trust each other. This should prevent an eavesdropper to
obtain the username and password, even in case of a man-in-the-middle
attack.