Add more limits to the number and frequency of submissions and usertests
Allow to set a per-contest and per-task limit on the total number of
submissions and usertest sent by each user and on their frequency.

The per-contest limit applies to *all* submissions or usertests of the
entire contest, while the per-task limit applies only to submissions or
usertest for that task.

Note that the default per-task limit of a submission every 60 seconds is
no more. It can be added back using AWS.