Remove some useless single_parent arguments from DB relationships
The single_parent argument for relationships is needed to make
many-to-one or many-to-many relationships become one-to-one or
one-to-many. It has no effect on relationships that already are of the
latter types. Yet, it was used on many backrefs which, usually, were
one-to-many relationships. These occurences have been removed.

The only remaining use is for tokens: it is needed to prevent
submissions from being associated with more than one token.