Add passive deletes
Since we have set a ondelete="CASCADE" option on all foreign keys we
don't need SQLAlchemy to *load* and explictly delete all children when a
parent is deleted (because the database will do that automatically).
The passive_delete option does exactly this.

The passive_update option is the same but fot updates. Yet, it's True by
default, so we don't need to set it explicitly.