Add some CheckConstraints in cms.db
We don't allow these values to be zero because setting a min_*_interval
to zero is the same as setting it to None and setting a max_*_number to
zero would disable submissions or usertest. While I think that someone
may want to disable them (well, at least the usertests...) I don't think
that the correct way to do that is to set the max_*_number to zero.
There should be a more explicit way to do that, to also allow the
interface (i.e. CWS) to know it and properly display this behavior.