OutputOnly task type look at parameter to decide the comparator.
Previously, the task type parameter for deciding if the task used a
white diff or a comparator was ignored and the decision taken from the
presence or absence of a manager with any name.

Now, the decision is taken looking at the parameter and we require the
manager to be called checker (to avoid hard-to-debug errors).