Fix dependecy lists
Add a few missing dependencies, use their "official" names (i.e. the one
they use on their website), add links to their homepages, update Ubuntu
installation command, add Arch Linux installation command. And do all of
this for both the list in the README and for the list in the
documentation.