cmstestsuite: Add simple tests for contest exporter and importer
Added support for running tests on an existing contest.
Use this to re-create the contest based on an export and run the tests again.

This is clearly not an exhaustive test (e.g. no idea if results are restored),
but does test whether tasks are recreated correctly.

A good next step would be to compare the results from both contests to ensure
they are identical.