Allow some leeway in wallclock timeout.
If very short timelimits are involved (<= 0.1s), a wallclock limit of 2*t isn't
always sufficient. For example, when communication tasks are involved, the user
executable may be blocked, waiting on the manager to initialise and open its
FIFO. Heavy disk I/O may also briefly stall the execution of a task without
giving it any CPU time.

Give it an extra second for cases like these.