TV: make AWS notify other services when Datasets or Testcases change
In particular, notify them when a Dataset is created, updated or deleted
and when Testcases are added or removed, since this could cause a change
in the ScoreTypes used by ScoringService (either in their name, their
parameters or the public testcases).

Also, when the set of the to-autojudge Datasets of a Task changes (when
a new Dataset is activated or when the autojudge flag is toggled),
notify both EvaluationService and ScoringService so they can check if
they find new submissions or user_tests that need to be processed.

This commit is part of the series that introduces task versioning