AWS: small fixes, whitespace and pseudo-PEP8
Not giving the parameter to the constructor of DB classes already sets
it to its default value: no need to be explicit.

Try to use a consistent style in most of the handlers: start by getting
the objects whose IDs are given by parameters, define self.contets as
soon as possible (as it could be used by the error handler), conclude
with try_commit, redirect differently based on its result (if
appropriate) and put the things "to do" between these two blocks,
separated by blank lines to highlight them.

Move all assignments for the "reevaluation_buttons" parameters together,
to highlight explictly where they're actually used (this becomes more
inefficient, since they're assigned many times with the same values, but
assignments are so simple that it shouldn't matter).

PEP8 states: "avoid more than one space around an assignment (or other)
operator to align it with another". I think this also applies to spaces
after commas in a list of tuples. And, anyway, what we had wasn't a real
alignment...

Remove an unused handler.