grading: some fixes to Job
Add the language to ExecutionJobs too (could be useful in the future for
interpreted languages), sort the fields to give them some sense (the
order should now be the one in which the fields are used/generated
during the compilation and evaluation process) and use an internal
namedtuple instead of abusing the Testcase DB class.

This improves the previous commit.