grading: fix a nasty subtle bug in the evaluation process
The whole issue lies in TaskType.py:237. It's iterating on the numbers
from zero (included) to len(testcases) (excluded) assuming that these
are the Testcases' nums. That's not true in general. Even if we're using
an OrderingList for dataset.testcases we have no guarantee that the nums
form a contiguos sequence (i.e. they don't have holes). For example,
deleting a Testcase via AWS doesn't load the OrderingList at all and
therefore doesn't preserve this property.

The test_number generated on that line is then given to the TaskType
that uses it as the num for the Evaluation, causing troubles for
ScoreTypes during the scoring phase because of mismatches with the
public_testcases.

To properly fix it I decided to make job.testcases become a dict with
int keys. This will not change the syntax used to access a testcase and
will thus make the existing code continue to work, but it will allow us
to give the correct test_number to TaskType.evaluate_testcase.

Note that this bug was NOT caused by the previous commit.