Remove UpdateDB
We decided that, starting with v1.1, when we release a new major version
we'll also release a script to update the dumps (that is, the contest
data exported by ContestExporter) from the format used by the previous
major version to the one used by the new one. This makes UpdateDB
useless, since it can be "simulated" by exporting the data, updating
CMS, dropping the DB to make CMS create the schema again and importing
the data again.