New IO manager based on gevent.
It should be a drop-in replacement for AsyncLibrary, although it
still requires proper testing, with the exception of threaded RPC,
which are not supported anymore (in a certain way, now all RPC are
"threaded", because they're executed in a dedicated greenlet; yet,
at the moment requesting a threaded RPC triggers an error).

So far there is not Web counterpart.