Migrate ScoringService to gevent.
This should be the last service, now Async is ready to be removed
from this branch. Yaeee!

On the negative side, SS was migrated to gevent, but it still retains
the horrible hack I put together for pushing to RWS using a
separate thread to prevent blocking. And let's avoid speaking of
the LogBridge hack. They'll have to be fixed as soon as possible,
now that we can.