Use JobGroup
Instead of combining the evaluations for all testcases into a single
EvaluationJob, use one for each testcase and introduce a JobGroup to
"bind" them all together.

We needed to remove the testcases and evaluations attributes from
EvaluationJob and add input, output and all the other fields that
TaskTypes dynamically put in the elements inserted in evaluations:
success, outcome, text, user_output, plus (user_output was called
output before but has been renamed to avoid collision).

JobGroup contains a dict of Jobs and a success attribute, which is
basically the AND of the success of all the Jobs it contains. The
methods to create Jobs from Submissions and UserTests have been moved
here too, with slight modifications (usually just the last two lines of
code, with the exception of from_submission_evaluation which required
some more changes).

The loop that breaks a JobGroup into single Jobs (the one that was in
TaskType.evaluate and called .evaluate_testcase) has been moved into
Worker. Therefore TaskType.evaluate has been removed and
.evaluate_testcase has been renamed to .evaluate. Worker, on the other
hand, doesn't need store the TaskType as class-level attribute just to
set its ignore_job attribute: it's now storing ignore_job itself (the
session attribute has also been removed as it was completely unused).