Identify Testcases with a string codename [2/3]
Mostly straightforward changes. Job.testcases and Job.evaluations are
now indexed by strings and not by integers (makes it also easier to
convert to and from JSON); the same holds for the public_testcases
argument for ScoreTypes; TaskType.evaluate_testcase also receives a
string. When EvaluationService creates an Evaluation it cannot pass the
codename to the constructor but has to use it to get the Testcase from
the Dataset.

When ScoreTypes need to put an order on Testcase (to group them into
subtasks or to present them in the graphical output) they use the
lexicographical order of the codenames.