Identify Testcases with a string codename [3/3]
Mostly straightforward again. Use the ID of the Testcase instead of its
codename ("namespaced" with the ID of the Dataset) as HTML form input
names (and hence HTTP query argument names) as it doesn't need to be
escaped and poses no risk. Given the new sorting criteria for
Evaluations in ScoreTypes, to preserve the current behavior import
Testcases with their numerical indexed (zero-padded to three digits) as
codename.