Implement stupid sandbox.
The stupid sandbox is an insecure sandbox, meant to be used for testing
CMS in environments where you can't use other sandboxes (for example
because of wrong kernel version, missing kernel options, lack of
superuser rights).

It must not be used in real contests. You can enable it setting the
hidden config option "sandbox_implementation" to "stupid".

This stupid sandbox is probably still a bit buggy. I hope, at least,
not to have introduced regressions in isolate.