SS: don't erase scores from RWS when active dataset is changed
Usually a dataset is activated only after it has been fully scored in
background. When that's not the case SS used to erase the scores of the
unscored submissions from RWS (just to push the new ones as soon as they
were available). After this commit scores that are readily available
will be pushed immediately, and the others will be kept unchanged until
they're scored again.