Fix bug in CWS
Sometimes downloading a file from CWS (for example a statement)
triggered a 500 HTTP error because a division by zero happened. I don't
know if it's because the timer had a low resolution or because the
transfer did really happen blazingly fast (or because the data was just
written to a buffer), but removing the speed computation surely fixes
the issue.