Fix a memory leak of SmartMappedCollection
cms.db.smartmappedcollection._event_manager is a global instance of
EventManager that keeps strong references to its handlers, which are
SmartMappedCollections that keep strong references to their parent
objects (i.e. the _parent_obj attribute). This prevents that object from
being garbage collected, and therefore its collection from being
unlinked, and therefore its handler from being removed. We got our
memory leak!

To fix it we just need to use weak references. And make the SMCs fit to
be elements of a set.