Fix untrusted input bug.
During the execution of a task of Batch type, the contestant program
could tamper with the input file (usually input.txt). We check it
out again from the database to be sure.

Committer note: this patch was first written when the contestant could
actually write the input file. At commit time, this is no longer true
(it can only write the output file). Nonetheless, graders might decide
to delete the input, so the patch still makes sense.