Be extra sure that no forbidden files are hanging around.
Hopefully this is not strictly required, but it is better to avoid
a blank catch-all except clause when dealing with security details.

The test may fail for example if the program manages to make the
directory not writable by the user running CMS. In this case the
removal fails, but CMS ignores it. I do not know if this could
happen now, but in line of principle it may happen in the future
if we change the owner model without remembering of this snippet.