Avoid errors for missing submission results when dealing with datasets
Creating a dataset without copying the results do not create empty
ones, but leave the submission without a result for the new
dataset. The code dealing with results needs to be made aware that
this might happen. In particular, this was not true for ProxyService
(when making the new dataset live) and for ES (when enabling
background judging).