Fix misuse of the Tornado API
This has long been a misuse: the second argument is 'strip', a boolean.
It has become critical with v4.2 because Tornado added an assert, to
"make sure `get_arguments` isn't accidentally being called with a
positional argument that's assumed to be a default (like in
`get_argument`.)". They caught us! :)