Fix some Session confusion in EvaluationService
Mainly, it was submission_enqueue_operations that was using Sessions
oddly. It received a Submission object, which usually has a session
already attached to it: the session that was used to fetch it from the
DB (this isn't true if the submission hasn't been stored yet, but it's
not our case). But then it was creating a new session, which it wasn't
using at all, since the only goal was to store it in extra, for check's
benefit.

The check function was called with either all parameters (in case of
submissions), which means both the session and all objects, or with none
of them (in case of user tests). In the former case it would just use
the objects and ignore the session, in the latter it would create its
own session and fetche the objects. Thus, in the case of submissions, it
wasn't actually checking again, as the same objects of the first check
were still being used. Moreover, this all means that the session created
in submission_enqueue_operations was never being used.

So now we have check fetch the objects again, instead of using the old
ones. And the session to do that is created by enqueue only if really
needed. Therefore submission_enqueue_operations doesn't have to pass any
extra anymore.