Remove JobGroup
Introduced in 967dc00, they were made useless by 7baeffe, that made it
possible for each testcase to be evaluated by a different worker.
Therefore a JobGroup always contained a single Job, making it
effectively superfluous.

The methods to build Jobs and to parse back their results into the DB
had to be moved into ESOperation, because the one for submission
evaluations required to know the testcase codename, which isn't stored
inside the Job itself and is thus unknown to it (with JobGroups, this
was available as the key of the dictionary). Basically, this means
reverting commit 916771f.