Refactor and pagination for submission lists and reevaluation in AWS
Implemented server side pagination so that the page in AWS and the
data transfered is not too huge. Also made explicit that global
reevaluation of contests and users includes submissions for all
datasets, not only the active ones.

There was also a bug that in single-submission reevaluation, the
reevaluation was always done on the active dataset and not on the
shown one.